Clean up misuse of types in buf_hash() function
* Use "unsigned long" for hash value instead of "long", to prevent the
  left shift causing signed overflow (which is undefined behaviour).
* Make all call sites store the return value as "unsigned long".
* Use "size_t" consistently for all string lengths.

+++ src/hl.c
@@ -38,9 +38,9 @@ static bool is_buffered(const struct condition *cond, const char *str, int len)
     return !memcmp(cond->u.cond_bufis.str, str, len);
 }
 
-static bool in_hash(struct string_list *list, const char *str, int len)
+static bool in_hash(struct string_list *list, const char *str, size_t len)
 {
-    unsigned int hash = buf_hash(str, len);
+    unsigned long hash = buf_hash(str, len);
     struct hash_str *h = list->hash[hash % ARRAY_COUNT(list->hash)];
 
     if (list->icase) {
